Let iOS use different audio load type based on clip length
This may be bad and/or unnecessary. The main tradeoff is that ADPCM is a bit larger than Vorbis and as I understand requires less compute to decompress than Vorbis. This commit generally will cost some CPU in favor of less memory (either compressing audio in memory or streaming directly from disk for larger clips).

Also adds loadInBackground to all clips, although this needs testing to see how it interacts with immediate playback.