Mobile Quality (#464)
* Add mobile renderer & URP asset

* Log full error on bundle skip

* Only pack normals if enabled; pack textures once

* Add iOS staging dev client build option w/ profiler

* Fix build menu function name

* Add lowest setting

Mainly for disabling shadows & variants associated

* auto create & configure quality levels

* configure quality on publish

* dont compile urp shaders when building core package

* swap quality levels when publishing packages

* Destroy VW meshes when deconstructed

* Let iOS use different audio load type based on clip length

This may be bad and/or unnecessary. The main tradeoff is that ADPCM is a bit larger than Vorbis and as I understand requires less compute to decompress than Vorbis. This commit generally will cost some CPU in favor of less memory (either compressing audio in memory or streaming directly from disk for larger clips).

Also adds loadInBackground to all clips, although this needs testing to see how it interacts with immediate playback.

* Only compress in memory on mobile; lower default audio quality

* quality level "low" instead of "mobile"

* Setup quality settings; don't free in deconstructor

* disable the new audio asset processor for now

---------

Co-authored-by: liamlutton <lutton+gh@me.com>